refactor: extract file operation services
